:recycle: Update FormsSerializer to TravelAuthorizationsSerializer.
