+++ api/src/serializers/index.ts
@@ -1,4 +1,4 @@
 export * from './expenses-serializer'
-export * from './forms-serializer'
+export * from './travel-authorizations-serializer'
 
 export default undefined
